Get parkinglot with division

+++ src/main/java/com/controlaltinsert/parkshark/division/domain/DivisionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DivisionRepository extends JpaRepository<Division,Long> {
+public interface DivisionRepository extends JpaRepository<Division, Integer> {
 
 }
